Stop truncating already-existing files.
Previously `experiment file` would truncate the file if the same identifier had
been used previously.  This was a mistake, since we never implemented the
badly-named --force.

Now, the file is created by opening it in append mode. We can implement
truncation later; since it is a dangerous option, it seems silly to make it the
default.